fix user logout; raise Unauthorized; note owner not nullable; add/edit notes

+++ app.py
@@ -1,7 +1,10 @@
 from flask import Flask, session, redirect, render_template, flash
 from flask_debugtoolbar import DebugToolbarExtension
+from werkzeug.exceptions import Unauthorized
+from werkzeug.wrappers import Response
 
-from models import db, connect_db, User
+
+from models import db, connect_db, User, Note
 
 from forms import RegistrationForm, LoginForm, CSRFProtectForm, NotesForm
 
@@ -73,19 +76,17 @@ def login_user():
         if user:
             session['username'] = user.username
             return redirect(f"/users/{user.username}")
-    else:
-        form.username.errors = ["Bad name/password"]
-        return render_template('login.html', form=form)
+        else:
+            form.username.errors = ["Bad name/password"]
+
+    return render_template('login.html', form=form)
 
 
 @app.get('/users/<username>')
 def display_user(username):
     """ displays user page """
-
     if SESSION_KEY not in session or session[SESSION_KEY] != username:
-        flash("You must be logged in to view!")
-        # TODO: raise Unauthorized error
-        return redirect("/")
+        raise Unauthorized()
 
     form = CSRFProtectForm()
     user = User.query.get_or_404(username)
@@ -101,7 +102,7 @@ def logout_user():
 
     if form.validate_on_submit():
         # Remove "user_id" if present, but no errors if it wasn't
-        session.pop("user_id", None)
+        session.pop("username", None)
 
     return redirect("/")
 
@@ -110,15 +111,69 @@ def delete_user(username):
     """deletes the provided user and all their notes. redirects to /"""
 
     user = User.query.get_or_404(username)
-    notes = Note.query.filter(username == user.username)
 
-    for note in notes:
-        db.session.delete(note)
+    Note.query.filter(username == user.username).delete()
 
     db.session.delete(user)
     db.session.commit()
 
     return redirect("/")
 
-@app.get("/users/<username>/notes/add")
-def add_note
+@app.route("/users/<username>/notes/add", methods=["GET", "POST"])
+def add_note(username):
+    """Add a note for the given user."""
+
+    if SESSION_KEY not in session or session[SESSION_KEY] != username:
+        raise Unauthorized()
+
+    form = NotesForm()
+    if form.validate_on_submit():
+        title = form.title.data
+        content = form.content.data
+
+        new_note = Note(title=title,content=content, owner_username=username)
+        db.session.add(new_note)
+        db.session.commit()
+        return redirect(f"/users/{username}")
+    else:
+        return render_template("add_note.html", form=form)
+
+
+@app.route("/notes/<note_id>/update", methods=["GET", "POST"])
+def edit_note(note_id):
+    """Edit a note for the given user."""
+
+    note = Note.query.get(note_id)
+    username = note.owner_username
+
+    if SESSION_KEY not in session or session[SESSION_KEY] != username:
+        raise Unauthorized()
+
+    form = NotesForm(obj=note)
+    if form.validate_on_submit():
+        note.title = form.title.data
+        note.content = form.content.data
+
+        db.session.commit()
+        return redirect(f"/users/{username}")
+
+    return render_template("edit_note.html", form=form)
+
+
+@app.post("/notes/<note_id>/delete")
+def delete_note(note_id):
+    """Delete a note."""
+
+    note = Note.query.get(note_id)
+    username = note.owner_username
+
+    if SESSION_KEY not in session or session[SESSION_KEY] != username:
+        raise Unauthorized()
+
+    form = CSRFProtectForm()
+
+    if form.validate_on_submit():
+        db.session.delete(note)
+        db.session.commit()
+
+    return redirect(f"/users/{username}")
